Removed phantom (duplicate) ship from the scene.
Reworked CameraManager, and ChaseCamera some.

Added Players node to pack tree, and simplified the codebase around players.

Players remain in the pack tree unless they specifically quit. It's basically the begining of a lobby.

Working on saving the last game scene image and fading back to the menu screen.

Starting camera for game play needs to be fixed.