Unifying spawning code and naming convention.
Adding LevelSettings.js to hold level specific settings. Unfortunately PlayCanvas does not allow for string arrays, or to reference objects in the scene in Script.Attributes, so the settings for each level will be hardcoded into the LevelSettings.js file.

Eventually LevelSettings.js will be attached to each root entity in each level, and the settings can be read from there. For the time being LevelSettings.js will be attached to the GameMenu.